cleaner fixes
