+++ apps/web/app/api/clean/save-result.ts
@@ -1,0 +1,45 @@
+import prisma from "@/utils/prisma";
+import { updateThread } from "@/utils/redis/clean";
+
+export async function saveCleanResult({
+  userId,
+  threadId,
+  archive,
+  jobId,
+}: {
+  userId: string;
+  threadId: string;
+  archive: boolean;
+  jobId: string;
+}) {
+  await Promise.all([
+    updateThread(userId, threadId, { status: "completed" }),
+    saveToDatabase({
+      userId,
+      threadId,
+      archive,
+      jobId,
+    }),
+  ]);
+}
+
+async function saveToDatabase({
+  userId,
+  threadId,
+  archive,
+  jobId,
+}: {
+  userId: string;
+  threadId: string;
+  archive: boolean;
+  jobId: string;
+}) {
+  await prisma.cleanupThread.create({
+    data: {
+      userId,
+      threadId,
+      archived: archive,
+      jobId,
+    },
+  });
+}
